examples as component
