+++ gradio/blocks.py
@@ -903,3 +903,5 @@ class Blocks(BlockContext):
             self.server.close()
             if self.enable_queue:
                 queueing.close()
+
+
